Unit Testing adjustments for clarity
