+++ Unit Testing with JUnit/src/calculator/model/calculator/Passenger.java
@@ -1,4 +1,4 @@
-package calculator.model.drive;
+package calculator.model.calculator;
 
 public class Passenger
 {
