--- Unit Testing with JUnit/src/calculator/tdd/CalculatorTDD.java
@@ -1,20 +1,0 @@
-package calculator.tdd;
-
-public class CalculatorTDD
-{
-    public int addNumbers(int num1, int num2){
-        return 0;
-    }
-
-    public int subtractNumbers(int num1, int num2){
-        return 0;
-    }
-
-    public int divideNumbers(int num1, int num2){
-        return 0;
-    }
-
-    public int multiplyNumbers(int num1, int num2){
-        return 0;
-    }
-}
