+++ Unit Testing with JUnit/test/calculator/model/CalculatorTest.java
@@ -1,0 +1,40 @@
+package calculator.model;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class CalculatorTest
+{
+
+    Calculator calculator = new Calculator();
+
+    @Test
+    public void addTest()
+    {
+
+    }
+
+
+    @Test
+    public void subtractTest()
+    {
+
+    }
+
+
+    @Test
+    public void multiplyTest()
+    {
+
+    }
+
+
+    @Test
+    public void divideTest()
+    {
+
+    }
+
+
+}
